Combine CQ Source and Source
