--- applications/source/geode-cq-source/README.adoc
@@ -1,30 +1,0 @@
-//tag::ref-doc[]
-= Geode Continuous Query Source
-
-The Geode CQ source will emit a stream of Objects extracted from Apache Geode CqEvents.
-
-== Options
-
-The **$$geode-cq$$** $$source$$ has the following options:
-
-//tag::configuration-properties[]
-$$geode.client.pdx-read-serialized$$:: $$Deserialize the Geode objects into PdxInstance instead of the domain class.$$ *($$Boolean$$, default: `$$false$$`)*
-$$geode.cq.supplier.event-expression$$:: $$SpEL expression to use to extract data from a cq event.$$ *($$Expression$$, default: `$$<none>$$`)*
-$$geode.cq.supplier.query$$:: $$The OQL query.$$ *($$String$$, default: `$$<none>$$`)*
-$$geode.pool.connect-type$$:: $$Specifies connection type: 'server' or 'locator'.$$ *($$ConnectType$$, default: `$$<none>$$`, possible values: `locator`,`server`)*
-$$geode.pool.host-addresses$$:: $$Specifies one or more Gemfire locator or server addresses formatted as [host]:[port].$$ *($$InetSocketAddress[]$$, default: `$$<none>$$`)*
-$$geode.pool.subscription-enabled$$:: $$Set to true to enable subscriptions for the client pool. Required to sync updates to the client cache.$$ *($$Boolean$$, default: `$$false$$`)*
-$$geode.region.region-name$$:: $$The region name.$$ *($$String$$, default: `$$<none>$$`)*
-$$geode.security.password$$:: $$The cache password.$$ *($$String$$, default: `$$<none>$$`)*
-$$geode.security.ssl.ciphers$$:: $$Configures the SSL ciphers used for secure Socket connections as an array of valid cipher names.$$ *($$String$$, default: `$$any$$`)*
-$$geode.security.ssl.keystore-type$$:: $$Identifies the type of Keystore used for SSL communications (e.g. JKS, PKCS11, etc.).$$ *($$String$$, default: `$$JKS$$`)*
-$$geode.security.ssl.keystore-uri$$:: $$Location of the pre-created Keystore URI to be used for connecting to the Geode cluster.$$ *($$Resource$$, default: `$$<none>$$`)*
-$$geode.security.ssl.ssl-keystore-password$$:: $$Password for accessing the keys truststore.$$ *($$String$$, default: `$$<none>$$`)*
-$$geode.security.ssl.ssl-truststore-password$$:: $$Password for accessing the trust store.$$ *($$String$$, default: `$$<none>$$`)*
-$$geode.security.ssl.truststore-type$$:: $$Identifies the type of truststore used for SSL communications (e.g. JKS, PKCS11, etc.).$$ *($$String$$, default: `$$JKS$$`)*
-$$geode.security.ssl.truststore-uri$$:: $$Location of the pre-created truststore URI to be used for connecting to the Geode cluster.$$ *($$Resource$$, default: `$$<none>$$`)*
-$$geode.security.ssl.user-home-directory$$:: $$Local directory to cache the truststore and keystore files downloaded form the truststoreUri and keystoreUri locations.$$ *($$String$$, default: `$$user.home$$`)*
-$$geode.security.username$$:: $$The cache username.$$ *($$String$$, default: `$$<none>$$`)*
-//end::configuration-properties[]
-
-//end::ref-doc[]
