--- applications/source/geode-cq-source/pom.xml
@@ -1,95 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>geode-cq-source</artifactId>
-    <version>3.0.0-SNAPSHOT</version>
-    <name>geode-cq-source</name>
-    <description>geode continuous query source app</description>
-    <packaging>jar</packaging>
-
-    <parent>
-        <groupId>org.springframework.cloud.stream.app</groupId>
-        <artifactId>stream-applications-core</artifactId>
-        <version>3.0.0-SNAPSHOT</version>
-        <relativePath/>
-    </parent>
-
-    <dependencies>
-
-        <dependency>
-            <groupId>org.springframework.cloud.fn</groupId>
-            <artifactId>geode-cq-supplier</artifactId>
-            <version>${java-functions.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-test</artifactId>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.projectlombok</groupId>
-            <artifactId>lombok</artifactId>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.cloud.fn</groupId>
-            <artifactId>function-test-support</artifactId>
-            <version>${java-functions.version}</version>
-            <scope>test</scope>
-        </dependency>
-
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.springframework.cloud</groupId>
-                <artifactId>spring-cloud-app-starter-doc-maven-plugin</artifactId>
-            </plugin>
-            <plugin>
-                <groupId>org.springframework.cloud.stream.app.plugin</groupId>
-                <artifactId>spring-cloud-stream-app-maven-plugin</artifactId>
-                <configuration>
-                    <generatedApp>
-                        <name>geode</name>
-                        <type>source</type>
-                        <version>${project.version}</version>
-                        <configClass>org.springframework.cloud.fn.supplier.geode.cq.GeodeCqSupplierConfiguration.class</configClass>
-                        <functionDefinition>geodeCqSupplier</functionDefinition>
-                    </generatedApp>
-                    <dependencies>
-                        <dependency>
-                            <groupId>org.springframework.cloud.fn</groupId>
-                            <artifactId>geode-cq-supplier</artifactId>
-                            <version>${java-functions.version}</version>
-                        </dependency>
-                    </dependencies>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-
-    <repositories>
-        <repository>
-            <snapshots>
-                <enabled>true</enabled>
-            </snapshots>
-            <id>spring-snapshots</id>
-            <name>Spring Snapshots</name>
-            <url>https://repo.spring.io/libs-snapshot-local</url>
-        </repository>
-        <repository>
-            <snapshots>
-                <enabled>false</enabled>
-            </snapshots>
-            <id>spring-milestones</id>
-            <name>Spring Milestones</name>
-            <url>https://repo.spring.io/libs-milestone-local</url>
-        </repository>
-    </repositories>
-
-</project>
