--- applications/source/geode-cq-source/src/main/resources/META-INF/dataflow-configuration-metadata-whitelist.properties
@@ -1,6 +1,0 @@
-configuration-properties.classes=org.springframework.cloud.fn.supplier.geode.cq.GeodeCqSupplierProperties,\
-org.springframework.cloud.fn.common.geode.GeodeClientCacheProperties,\
-org.springframework.cloud.fn.common.geode.GeodeRegionProperties,\
-org.springframework.cloud.fn.common.geode.GeodePoolProperties,\
-org.springframework.cloud.fn.common.geode.GeodeSecurityProperties,\
-org.springframework.cloud.fn.common.geode.GeodeSslProperties
