+++ applications/source/pom.xml
@@ -11,7 +11,6 @@
 
     <modules>
         <module>geode-source</module>
-        <module>geode-cq-source</module>
         <module>file-source</module>
         <module>jdbc-source</module>
         <module>http-source</module>
