--- functions/supplier/geode-cq-supplier/pom.xml
@@ -1,79 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>geode-cq-supplier</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
-    <name>geode-cq-supplier</name>
-    <description>geode continuous query supplier</description>
-
-    <parent>
-        <groupId>org.springframework.cloud.fn</groupId>
-        <artifactId>spring-functions-parent</artifactId>
-        <version>1.0.0-SNAPSHOT</version>
-        <relativePath>../../spring-functions-parent</relativePath>
-    </parent>
-    <properties>
-        <spring-cloud-fn.version>${project.version}</spring-cloud-fn.version>
-    </properties>
-    <dependencies>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.integration</groupId>
-            <artifactId>spring-integration-gemfire</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.cloud.fn</groupId>
-            <artifactId>geode-common</artifactId>
-            <version>${spring-cloud-fn.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.cloud.fn</groupId>
-            <artifactId>config-common</artifactId>
-            <version>${spring-cloud-fn.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.hibernate.validator</groupId>
-            <artifactId>hibernate-validator</artifactId>
-            <optional>true</optional>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-configuration-processor</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-test</artifactId>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.cloud.fn</groupId>
-            <artifactId>function-test-support</artifactId>
-            <version>${spring-cloud-fn.version}</version>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.projectlombok</groupId>
-            <artifactId>lombok</artifactId>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>io.projectreactor</groupId>
-            <artifactId>reactor-test</artifactId>
-		</dependency>
-
-	</dependencies>
-
-</project>
