--- functions/supplier/geode-cq-supplier/src/main/resources/geode-client.properties
@@ -1,4 +1,0 @@
-#
-# Geode client pool must be configured with subscriptions enabled for this component.
-#
-geode.pool.subscriptionEnabled=true
