+++ functions/supplier/geode-supplier/README.adoc
@@ -2,24 +2,36 @@
 
 This module provides a `java.util.function.Supplier` that can be reused and composed in other applications.
 The `Supplier` configures an Apache Geode client that connects to an external Apache Geode cache server or locator to monitor an existing region.
-The supplier uses the `CacheListeningMessageProducer` from `Spring Integration` which wraps a https://geode.apache.org/releases/latest/javadoc/org/apache/geode/cache/CacheListener.html[CacheListener]
+If `geode.supplier.query` is provided, the supplier will create a continuous query on the region and publish any events that meet the select criteria.
+If no query is provided, the supplier will publish all create and update events on the region.
+
+## Continuous Query
+If a query is provided, the supplier uses the `ContinuousQueryMessageProducer` from `Spring Integration` which wraps a https://docs.spring.io/spring-data/gemfire/docs/current/api/org/springframework/data/gemfire/listener/ContinuousQueryListenerContainer.html[ContinuousQueryListenerContainer]
+and emits a reactive stream of objects, extracted from a Geode https://geode.apache.org/releases/latest/javadoc/org/apache/geode/cache/CqEvent.html[CqEvent], which holds all of the
+event details.
+A SpEl Expression, given by the property `geode.cq.supplier.event-expression` is used to extract desired fields from the CQEvent payload.
+The default expression is `newValue` which returns the current value from the configured Region.
+
+## Cache Listener
+
+If no query is provided, the supplier uses the `CacheListeningMessageProducer` from `Spring Integration` which wraps a https://geode.apache.org/releases/latest/javadoc/org/apache/geode/cache/CacheListener.html[CacheListener]
 and emits a reactive stream of objects, extracted from a Geode https://geode.apache.org/releases/latest/javadoc/org/apache/geode/cache/EntryEvent.html[EntryEvent], which holds all of the
 event details.
 A SpEl Expression, given by the property `geode.supplier.entry-event-expression` is used to extract desired fields from the EntryEvent payload.
 The default expression is `newValue` which returns the current value from the configured Region.
 
 NOTE: Retrieving the value by itself is not always sufficient, especially if it does not contain the key value, or any additional context.
-The key is referenced by the field `key`. The EntryEvent also provides the `operation` (create or update).
+The key is referenced by the field `key`.
 If the cached key and value types are primitives, a simple expression like `key + ':' +newValue` may be useful.
 To access the entire EntryEvent, set the expression to `#root` or `#this`.
 
-The CacheListeningMessageProducer emits objects to the supplier implemented as `Supplier<Flux<?>>`.
+The configured MessageProducer emits objects to the supplier implemented as `Supplier<Flux<?>>`.
 Users have to subscribe to the returned `Flux` to receive the data.
 
 ## PDX Serialization
 
-The supplier works with PDX serialized cache objects of type https://geode.apache.org/releases/latest/javadoc/org/apache/geode/pdx/PdxInstance.html[PdxInstance], which Geode uses to store objects that can be represented as JSON. If the target region uses PDX serialization and you set  `geode.client.pdx-read-serialized` to `true`, PdxInstance objects will be returned as JSON strings.
-
+The supplier works with PDX serialized cache objects of type https://geode.apache.org/releases/latest/javadoc/org/apache/geode/pdx/PdxInstance.html[PdxInstance], which Geode uses to store objects that can be represented as JSON.
+If the target region uses PDX serialization and you set  `geode.client.pdx-read-serialized` to `true`, PdxInstance objects will be returned as JSON strings.
 
 ## Beans for injection
